+++ RELEASES.md
@@ -939,3 +939,13 @@ Based on:
 - [typescript v0.6.0-alpha.2] .
 ### Releases
 - [NPM v0.6.0-alpha.2] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.2 - .
+
+## 2026-02-06 02:13:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.704.0 (2.806.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0-alpha.3] .
+### Releases
+- [NPM v0.6.0-alpha.3] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.3 - .